MWW: fixed some bugs related to the ASTMapVisitor: first it wasn't doing the right thing for local equations (an unimplemented method).  Second, the AgreeProgram AST uses AgreeNode references to manage connections, which means that AgreeProgram is an ASG (as in graph) rather than an AST (as in tree).  This makes it somewhat more dangerous to transform, and one of the earlier passes was not maintaining these links correctly.
Rather than fix the offending pass, I made the ASTMapVisitor more robust:
it now uses the underlying ComponentInstance to do the connection mapping
after a transformation rather than the AgreeNode pointer, which may change
during the transformation.

Also added a prettyprinter.  The other changes may have been only to
add/remove printlns to get a handle on what was going on with the
ASTMapVisitor.