Add support for AADL arrays and the 'forall' AGREE keyword
The commit builds the framework needed for the back-end implementation
of the array extensions in the AGREE syntax and implements in that
framework initial support for AADL arrays and the 'forall' keyword.

If the array is declared over a range constrained data-type, then the
normal assertions for the range on the type will not appear in the
final LUSTRE model. In general arrays over data-types with range
constraints are not yet support.

Adding in the implementations for the 'exists' and 'foreach' keywords
should be relatively straight forward following the pattern in the
'forall' implementation.

AADL arrays have been tested. AGREE arrays have not been tested. There
are several example files that illustrate the AADL arrays.